+++ apps/qbittorrent-cleanup/base/kustomization.yaml
@@ -2,7 +2,7 @@
 resources:
   - cronjob.yaml
 images:
-  - name: ghcr.io/shikanime/manifests/qbittorrent-resume
+  - name: ghcr.io/shikanime/manifests/qbittorrent-cleanup
     digest: sha256:91450334c9b119d94bd98c3d70f5025e271f1d356c4388446f9556121588c918
-    newName: ghcr.io/shikanime/manifests/qbittorrent-resume
+    newName: ghcr.io/shikanime/manifests/qbittorrent-cleanup
     newTag: c4f877de1
