+++ clusters/nishir/overlays/tailnet/ks.yaml
@@ -534,17 +534,17 @@ spec:
 apiVersion: kustomize.toolkit.fluxcd.io/v1
 kind: Kustomization
 metadata:
-  name: apps-qbittorrent-resume
+  name: apps-qbittorrent-cleanup
   namespace: flux-system
 spec:
   dependsOn:
     - name: infrastructure-cert-manager
     - name: infrastructure-vpa
   interval: 10m
-  path: ./apps/qbittorrent-resume/overlays/nishir-tailnet
+  path: ./apps/qbittorrent-cleanup/overlays/nishir-tailnet
   prune: true
   healthChecks:
-    - name: qbittorrent-resume
+    - name: qbittorrent-cleanup
       apiVersion: batch/v1
       kind: CronJob
       namespace: shikanime
